[Core/Test] Break down `test_ray_init.py` by 2 files (#27635)
Seems like this test starts taking longer than 300 seconds because of newly added tests. Breaking down tests into 2 files
